# docs/API.openapi.yaml
openapi: 3.0.0
info:
  title: BuildBrainOS API
  description: |
    Complete REST API specification for BuildBrainOS platform.
    All endpoints require JWT authentication in Authorization header.
    
    Base URLs:
    - Development: http://localhost:4000/api/v1
    - Production: https://api.buildbrainos.com/api/v1
    
    Authentication:
    - JWT Bearer Token in Authorization header
    - Refresh tokens valid for 30 days
    - Rate limiting: 1000 requests/hour per user
  version: 1.0.0
  contact:
    name: BuildBrainOS Support
    email: support@buildbrainos.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:4000/api/v1
    description: Development server
  - url: https://api.buildbrainos.com/api/v1
    description: Production server
...
paths:
...
  /projects:
    post:
      tags:
        - Projects
      summary: Create new project
      operationId: createProject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - location
                - budget
                - timeline
              properties:
                title:
                  type: string
                  example: Downtown Office Complex - Phase 2
                description:
                  type: string
                location:
                  $ref: '#/components/schemas/Address'
                budget:
                  type: number
                  format: double
                  example: 1200000
                timeline:
                  type: object
                  properties:
                    startDate:
                      type: string
                      format: date
                    endDate:
                      type: string
                      format: date
                scope:
                  type: array
                  items:
                    type: string
                  example: [Framing, Electrical, Plumbing]
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'

    get:
      tags:
        - Projects
      summary: List user's projects
      operationId: listProjects
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, COMPLETED, PAUSED, ARCHIVED]
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  total:
                    type: integer
                  skip:
                    type: integer
                  limit:
                    type: integer
  /projects/{projectId}:
    get:
      tags:
        - Projects
      summary: Get project details
      operationId: getProject
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'

    patch:
      tags:
        - Projects
      summary: Update project
      operationId: updateProject
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                status:
                  type: string
                  enum: [ACTIVE, COMPLETED, PAUSED, ARCHIVED]
                budget:
                  type: number
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /projects/{projectId}/milestones:
    post:
      tags:
        - Projects
      summary: Create project milestone
      operationId: createMilestone
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - amount
                - dueDate
              properties:
                title:
                  type: string
                  example: Foundation Complete
                description:
                  type: string
                amount:
                  type: number
                  format: double
                dueDate:
                  type: string
                  format: date
                deliverables:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Milestone created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Milestone'

    get:
      tags:
        - Projects
      summary: List project milestones
      operationId: listMilestones
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Milestone'
...
  /payments/ledger:
    post:
      tags:
        - Payments
      summary: Request payment approval
      operationId: requestPayment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - projectId
                - milestoneId
                - amount
              properties:
                projectId:
                  type: string
                milestoneId:
                  type: string
                amount:
                  type: number
                  format: double
                description:
                  type: string
                documentProof:
                  type: array
                  items:
                    type: string
                    format: uri
                  description: Photo/document URLs for verification
      responses:
        '201':
          description: Payment request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
  /payments/ledger/{paymentId}:
    get:
      tags:
        - Payments
      summary: Get payment details
      operationId: getPayment
      security:
        - bearerAuth: []
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'

    patch:
      tags:
        - Payments
      summary: Update payment status
      operationId: updatePaymentStatus
      security:
        - bearerAuth: []
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [APPROVED, REJECTED, CANCELLED]
                note:
                  type: string
      responses:
        '200':
          description: Payment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
  /payments/wallet:
    get:
      tags:
        - Payments
      summary: Get user wallet balance
      operationId: getWallet
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Wallet information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
  /payments/withdraw:
    post:
      tags:
        - Payments
      summary: Request fund withdrawal
      operationId: requestWithdrawal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
                - method
              properties:
                amount:
                  type: number
                  format: double
                method:
                  type: string
                  enum: [ACH, WIRE, CARD]
                bankAccount:
                  type: object
                  properties:
                    accountNumber:
                      type: string
                    routingNumber:
                      type: string
                    accountType:
                      type: string
                      enum: [CHECKING, SAVINGS]
      responses:
        '201':
          description: Withdrawal initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  withdrawalId:
                    type: string
                  status:
                    type: string
                  estimatedArrival:
                    type: string
                    format: date
  /documents/upload:
    post:
      tags:
        - Documents
      summary: Upload project document
      operationId: uploadDocument
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - type
              properties:
                file:
                  type: string
                  format: binary
                type:
                  type: string
                  enum: [CONTRACT, LIEN_WAIVER, INSURANCE, LICENSE, BLUEPRINT, PERMIT, RFI]
                projectId:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
...
components:
  schemas:
...
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        country:
          type: string
        latitude:
          type: number
        longitude:
          type: number
    Project:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [ACTIVE, COMPLETED, PAUSED, ARCHIVED]
        budget:
          type: number
          format: double
        spent:
          type: number
          format: double
        location:
          $ref: '#/components/schemas/Address'
        timeline:
          type: object
          properties:
            startDate:
              type: string
              format: date
            endDate:
              type: string
              format: date
            progressPercent:
              type: number
        scope:
          type: array
          items:
            type: string
        gc:
          type: object
          properties:
            id:
              type: string
            companyName:
              type: string
        milestones:
          type: array
          items:
            $ref: '#/components/schemas/Milestone'
        createdAt:
          type: string
          format: date-time
    Milestone:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, COMPLETED, APPROVED, REJECTED]
        amount:
          type: number
          format: double
        dueDate:
          type: string
          format: date
        deliverables:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
...
    Payment:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        milestoneId:
          type: string
        senderId:
          type: string
        recipientId:
          type: string
        amount:
          type: number
          format: double
        status:
          type: string
          enum: [PENDING, APPROVED, PROCESSING, COMPLETED, FAILED, CANCELLED]
        method:
          type: string
          enum: [ACH, CARD, INTERNAL_LEDGER, USDC, WIRE]
        externalTransactionId:
          type: string
        verificationStatus:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    Wallet:
      type: object
      properties:
        userId:
          type: string
        usdBalance:
          type: number
          format: double
        usdcBalance:
          type: number
          format: double
        platformCredits:
          type: number
          format: double
        kycTierLimit:
          type: number
          format: double
        monthlyFeelingLimit:
          type: number
          format: double
        monthlySettledAmount:
          type: number
          format: double
        pendingPayments:
          type: number
          format: double
        lastUpdated:
          type: string
          format: date-time
    Document:
      type: object
      properties:
        id:
          type: string
        fileName:
          type: string
        type:
          type: string
          enum: [CONTRACT, LIEN_WAIVER, INSURANCE, LICENSE, BLUEPRINT, PERMIT, RFI]
        url:
          type: string
          format: uri
        status:
          type: string
          enum: [UPLOADED, PROCESSING, VERIFIED, FAILED]
        extractedData:
          type: object
        aiVerified:
          type: boolean
        uploadedAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in Authorization header
...